requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.41
aiosqlite==0.21.0
python-jose[cryptography]==3.5.0
passlib==1.7.4
python-multipart==0.0.20
pydantic-settings==2.10.1
greenlet==3.2.4
